Approve requests

+++ auto.sh
@@ -11,8 +11,8 @@ STACK_NAME="SlackGroupBot"
 BUILD_ARTIFACT_BUCKET="dev-lightraildevartifacts-ywjp7wt8djk7-bucket-1mlnqtwvk2jzf"
 
 # Parameter values for the sam template.  see: `aws cloudformation deploy help`
-PARAMETER_OVERRIDES=""
-#PARAMETER_OVERRIDES="--parameter-overrides KeyOne=value KeyTwo=value"
+#PARAMETER_OVERRIDES=""
+PARAMETER_OVERRIDES='--parameter-overrides GroupBotProject=Groot Accounts={"dev":"757264843183"} SlackToken=Jsy6HbHkV3ph7VQjuebXOQEA'
 
 USAGE="usage: $0 <command name>\nvalid command names: build delete deploy invoke upload"
 
@@ -66,7 +66,15 @@ elif [ "$COMMAND" = "deploy" ]; then
     aws cloudformation package --template-file infrastructure/$TEMPLATE.yaml --s3-bucket $BUILD_ARTIFACT_BUCKET --output-template-file "$OUTPUT_TEMPLATE_FILE"
 
     echo "Executing aws cloudformation deploy..."
-    aws cloudformation deploy --template-file "$OUTPUT_TEMPLATE_FILE" --stack-name $STACK_NAME --capabilities CAPABILITY_IAM $PARAMETER_OVERRIDES
+    aws cloudformation deploy --template-file "$OUTPUT_TEMPLATE_FILE" --stack-name $STACK_NAME --capabilities CAPABILITY_NAMED_IAM $PARAMETER_OVERRIDES
+
+    if [ $? -ne 0 ]; then
+        # Print some help on why it failed.
+        echo ""
+        echo "Printing recent CloudFormation errors..."
+        aws cloudformation describe-stack-events --stack-name $STACK_NAME --query 'reverse(StackEvents[?ResourceStatus==`CREATE_FAILED`||ResourceStatus==`UPDATE_FAILED`].[ResourceType,LogicalResourceId,ResourceStatusReason])' --output text
+        exit 4
+    fi
 
     # cleanup
     rm "$OUTPUT_TEMPLATE_FILE"
